221202 - store 디렉토리 생성
