--- README.md
@@ -1,2 +1,0 @@
-# To-do-list-app
-✔️ 바닐라 자바스크립트 To Do List App
