+++ index.html
@@ -1,0 +1,78 @@
+<!DOCTYPE html>
+<html lang="ko">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <!-- LINK -->
+  <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/reset-css@5.0.1/reset.min.css">
+  <link rel="stylesheet" href="./css/main.css">
+  <!-- GOOGLE FONTS -->
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Black+Han+Sans&family=Erica+One&display=swap" rel="stylesheet">
+  <!-- SCRIPT -->
+  <script defer type="module" src="./js/main.js"></script>
+  <title>To Do List App</title>
+</head>
+
+<body>
+  <main>
+    <div class="container">
+      <div class="title">
+        <h1>To Do List</h1>
+      </div>
+      <div class="write">
+        <input type="text" id="todo-input" class="todo-input" placeholder="할 일을 추가하세요." />
+        <button type="button" id="todo-btn" class="todo-btn">Add</button>
+      </div>
+      <div class="list" id="list">
+        <div class="list-date" id="list-date"></div>
+        <ul class="list-wrap" id="list-wrap">
+          <!-- item -->
+        </ul>
+        <img src="./images/loading.gif" class="loading" alt="">
+        <!-- <div class="list-btn">
+          <button id="list-delete">DELETE</button>
+        </div>
+        <div class="complete">
+          <p>2/3</p>
+        </div> -->
+      </div>
+      <div class="button">
+        <!-- <button>NEW</button> -->
+        <!-- <button>All DELETE</button> -->
+      </div>
+    </div>
+  </main>
+  <section class="calendar">
+    <div class="title">
+      <h2>Calendar</h2>
+    </div>
+    <div class="body">
+      <div class="container">
+        <div class="month-wrap">
+          <button class="prev"></button>
+          <div class="month"></div>
+          <button class="next"></button>
+        </div>
+        <button class="today">Today</button>
+        <div class="days-wrap">
+          <div class="days">
+            <div class="day">S</div>
+            <div class="day">M</div>
+            <div class="day">T</div>
+            <div class="day">W</div>
+            <div class="day">T</div>
+            <div class="day">F</div>
+            <div class="day">S</div>
+          </div>
+          <div class="dates"></div>
+        </div>
+      </div>
+    </div>
+  </section>
+</body>
+
+</html>
