+++ public/main.css
@@ -1,0 +1,2 @@
+h1 {
+  color: red; }
