--- to-do-app/js/createTodo.js
@@ -1,19 +1,0 @@
-export default async function createTodo(todoEl, changeToSix) {
-  console.log('createTodo' + changeToSix);
-  const res = await fetch('https://asia-northeast3-heropy-api.cloudfunctions.net/api/todos', {
-    method: 'POST',
-    headers: {
-      'content-type': 'application/json',
-      'apikey': 'FcKdtJs202209',
-      'username': 'KDT3_WooJiSoo'
-    },
-    body: JSON.stringify({
-      title: `${todoEl}`,
-      order: changeToSix,
-    })
-  });
-  const todos = await res.json();
-  console.log(todos.title);
-  console.log(todos);
-  return todos;
-}
