--- to-do-app/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "vite-project",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "devDependencies": {
-    "sass": "^1.56.1",
-    "vite": "^3.2.3"
-  }
-}
