--- to-do-app/public/main.css
@@ -1,2 +1,0 @@
-h1 {
-  color: red; }
